connection issue fixed

+++ Mingems.Infrastructure/Services/InvestmentService.cs
@@ -47,6 +47,9 @@ namespace Mingems.Infrastructure.Services
             var investment = await unitOfWork.InvestmentRepository.GetByIdAsync(Id);
             if (investment == null)
                 throw new NotFoundException("Investment not found or already removed");
+            var purchases = await unitOfWork.PurchaseRepository.GetAllAsync();
+            if (purchases.Where(p => p.InvestorId == investment.Id) != null)
+                throw new InvalidException($"{investment.FirstName} is already connected with another record, please remove the connected records!");
             unitOfWork.InvestmentRepository.Remove(investment.Delete(email));
             await unitOfWork.CommitAsync();
         }
